Move /journal onto the unified Document model (PR 4/4)
Journal list, Calendar, and Heatmap now read from the personal
space's JOURNAL documents instead of the legacy journal_entry
table. /spaces and /journal are now two views into the same
document model, with personal space hidden from the spaces list
and surfaced exclusively as /journal.

- GET /api/spaces/personal: returns the user's auto-created
  personal space (every user has one, courtesy of V4 backfill)
- GET /api/spaces/{id}/documents/calendar: JOURNAL docs with
  entry_date; hasImage computed in one query, not N+1
- GET /api/spaces/{id}/documents/heatmap?year=Y: aggregate
  count-per-day via JPQL GROUP BY
- listDocuments now accepts an optional date= filter so the
  Calendar's date-click drill-in still works against the new model
- JournalListPage rebuilt around the doc cards; "New entry"
  navigates straight to the editor. Legacy keyword/AI search/
  recap/prompts UI is removed for now — those endpoints still
  hit /api/entries and will migrate in a follow-up
- CalendarEventResponse.id Integer -> Long to fit document IDs

+++ src/main/java/com/myjourney/dto/CalendarEventResponse.java
@@ -5,7 +5,9 @@ package com.myjourney.dto;
  * `hasImage` lets the UI render a small camera glyph in the event pill when true.
  */
 public record CalendarEventResponse(
-        Integer id,
+        // Long covers both legacy journal_entry IDs (INT, widened) and the
+        // new document IDs (BIGINT). Frontend deserializes either as `number`.
+        Long id,
         String title,
         String start,
         boolean hasImage
